Users can no longer delete files via browser address bar, unless they have privileges into the folder

+++ app/Http/Controllers/FileController.php
@@ -9,6 +9,7 @@ use Filebrowser\Http\Controllers\Controller;
 use Illuminate\Support\Facades\View;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
+use Filebrowser\Folder;
 use app\Download;
 use Storage;
 
@@ -60,18 +61,30 @@ public function upload(Request $request){
 public function delete($file){
   //Get the current logged in user's name
   $userName = Auth::user()->name;
+  //Get the current user ID
+  $userID = Auth::user()->id;
   //Explode the filepath
   $fileExpl = explode("/", $file);
-  //Use the last element in exploded array
-  $fileName = end($fileExpl);
-  //Remove the filename from received filepath to get the directory
-  $filepath = str_replace($fileName, "", $file);
-  //Deletes the file that it received.
-  Storage::delete($file);
-  //Insert activity
-  app('Filebrowser\Http\Controllers\ActivityController')->updateActivityLog($userName, "deleted", "file", $fileName, "from", $filepath);
-  //Updates page and shows notification about the successful deleton.
-  return back()->with('delete', 'Tiedoston poisto onnistui');
+  //Get the folder
+  $folder = Folder::where('folder_name', '=' , $fileExpl[1])->first();
+  //Get the folder's ID
+  $folderID = $folder->id;
+  //Check if current user has access into a selected directory
+  if(DB::table('folder_user')->where('folder_id', '=', $folderID)->where('user_id', '=', $userID)->exists()){
+    //Use the last element in exploded array
+    $fileName = end($fileExpl);
+    //Remove the filename from received filepath to get the directory
+    $filepath = str_replace($fileName, "", $file);
+    //Deletes the file that it received.
+    Storage::delete($file);
+    //Insert activity
+    app('Filebrowser\Http\Controllers\ActivityController')->updateActivityLog($userName, "deleted", "file", $fileName, "from", $filepath);
+    //Updates page and shows notification about the successful deleton.
+    return back()->with('delete', 'Tiedoston poisto onnistui');
+  }
+  else{
+    return redirect("/")->with('error', 'You do not have access to this directory');
+  }
 }
 
 public function rename(Request $request){
